firmware: ipq-wifi: support 160mhz on ZyXEL NBG7815 (2025-06-13)
Upstream is using 2.7.0.1 board-2-bin version for id=255. Zyxel nbg7815
has this board id.

In order to have one 5ghz 160mhz like the factory firmware it is needed
to replace original board2.bin with this one.

qualcommax: nbg7815: extract only 2 mac for wifi

Now that we have 160mhz support for 5ghz band, we dont need 3 MAC for
wifi because we have only 2 radios.